+++ NEWS.md
@@ -1,5 +1,10 @@
 # JShotgun news
 
+## 1.1.0 / 2014-08-29
+
+- Added ShotgunFilter for web frameworks that use Filters instead of Servlets
+- Separate Shotgun logic out so it can be used for non-servlet framework code
+
 ## 1.0.1 / 2013-07-29
 
 - Fix double loading of classes. #1